improve #116 - add weight-score for X/Y

+++ kwm/window.h
@@ -54,7 +54,7 @@ void MoveCursorToCenterOfFocusedWindow();
 void ShiftWindowFocus(int Shift);
 void ShiftWindowFocusDirected(int Degrees);
 window_info FindClosestWindow(int Degrees);
-int GetWindowDistance(window_info *A, window_info *B);
+double GetWindowDistance(window_info *A, window_info *B);
 void GetCenterOfWindow(window_info *Window, int *X, int *Y);
 bool WindowIsInDirection(window_info *A, window_info *B, int Degrees);
